WIP factor out positioning wrapper
